Delete the webpage and call exit on the event loop.
This is due to a change in behaviour in QtWebkit 2.2. Now what happens is that there are 2 loadFinished()
signals emitted, from the QWebPage and from the QWebPage's QWebFrame. If we listen to the QWebPage's loadFinished()
and a script calls phantom.exit() in the callback, the QWebPage/QWebFrame will be deleted before the second signal is
emitted, which will crash.

One option would be to listen to the QWebFrame loadFinished signal instead, but then we rely on the order of signals being
emitted, which is really fragile. In addition, it still crashes because there is some cleanup being done in the destructors of
the QtWebKit's QNetworkReply that assumes that the QWebPage/QWebFrame objects still exist.

This bug is reported to the QtWebkit guys, so this is a temporary fix