Fix file upload on Qt 4.8.
Code was added to WebKit to detect whether the click event on the file
input element is triggered by user action or by a script. If it's by a
script, the event is ignored. This prevented us from simulating a click
on the element.

This fix just patches WebKit to remove the detection.

The relevant WebKit commit is:

  https://github.com/WebKit/webkit/commit/931870c11b2a4d0b714eea241db5654d9d8cf6a7

The associated WebKit bug is:

  https://bugs.webkit.org/show_bug.cgi?id=47593

Our bug is:

  http://code.google.com/p/phantomjs/issues/detail?id=307&q=upload

Conflicts:

	deploy/build-mac.sh