server: make it possible to access POST/PUT data
For POST and PUT request we now read all data as defined by
the Content-Length header into request.rawData property.
This property is a QByteArray which neatly maps to an array
in javascript.

For POST requests with Content-Type = applicaiton/x-www-form-urlencoded
we furthermore provide a parsed, easy-to-use request.post property.
This one is a QVariantMap of the decoded form data.

The simpleserver example is extended to get a simple form with
a submit button and text field. The latter also shows that e.g.
UTF-8 data is handled as one would expect.