Draw SVG lines with stroke-opacity in PDF output.
Webkit tells the PDF painter about SVG stroke opacity by setting the
global painter opacity state, *not* the pen (line drawer)'s color's
alpha.  So if the pen was opaque, but the painter-global opacity was
not, line drawing would choose the wrong path (one of the "simplePen"
paths) in QPdfBaseEngine.

We simply ensure that global opacity is considered when choosing
whether the "simplePen" path is appropriate or not.

https://github.com/ariya/phantomjs/issues/11381