automatic screenshot size based on document size
this change allows that (when no size was specified) an attempt to detect the size automatically based on the document is done,
first we try to use the `body` size, if no `body` is found, just uses the document root (eg. an `svg`).
If this fails, it fallbacks to current behavior (ie. 600x600).
This allows to take screenshot of html where the size depends on its content.