Update 2000-01-07-related-projects.md
I found the way I was using PhantomJS nice and useful for other people, so I've decided to take a little time, clean up and document the code and open source it. Basically I've built a NodeJS server that starts multiple PhantomJS instances and communicates through HTTP with them (using the built-in Mongoose).

I wasn't really sure under which category to place it: "In the family" or "Tools"?