Give a IntPoint to paint instead of x and y
This fixes the last errors after 218a937.

I have confirmed that both the thead and the tfoot blocks are repeated
on the next pages when a table breaks over multiple pages. The other
content was already moved to take account of the size of thead and
tfoot, so this change doesn't affect any other content.

This fixes issue #13324.
https://github.com/ariya/phantomjs/issues/13324