Repeat thead and tfoot when table contains page breaks
This has been done a few times before[0][1], but seems to be broken
again as reported in [2].

This commit is only reinstating what's missing from [1]. This was done
by running:

    git log -p -1 2d778f6 | git apply -3 -p6 --directory=src/qt/qtwebkit

And resolving the conflicts. There are still some compile errors. These
will be fixed in the upcoming commits instead of in this one to clarify
the changes.

[0]: https://github.com/ariya/phantomjs/pull/211
[1]: https://github.com/ariya/phantomjs/pull/344
[2]: https://github.com/ariya/phantomjs/issues/13324