Add ability to test exit status and stdin/out/err directly.
This works with more //! directives at the top of a test, which
will normally be a no-harness test.  It is now possible, for
instance, to test for an _unsuccessful_ exit, or for output
appearing on stderr rather than stdout.  The tests for
phantom.exit and sys.stdin/out/err have been augmented
accordingly.

It is also now possible to adjust the backstop timeout on a
per-test basis, and therefore the existing timeouts have been
lowered to 5s (harness), 7s (backstop).

Part of issue #13478 (test suite overhaul).