Fix compilation with GCC 5. (Issue #13518)
When an out-of-line template is used outside the file it's defined in, the
programmer is obliged to explicitly list (in the file with the definition)
the set of instantiations required by other files, even if all of those
instantiations are also required by the file with the definition.  GCC 5
inlines a template instantiation that 4.x didn't, causing a use from another
file to go unsatisfied and the link to fail.

Patch by Khem Raj, as applied in upstream Webkit.
Upstream Qt bug: https://bugreports.qt.io/browse/QTBUG-44829
Upstream Webkit bug: https://bugs.webkit.org/show_bug.cgi?id=147815