Fix for issue #10455: Canvas toDataURL encode is different.
Prior to the fix, quality would default to 100 irrespective of the image type, if an explicit quality-argument was not provided, or if the quality was outside the valid range. In the case where toDataURL is called without any arguments, the image type defaults to `image/png`, and quality defaults to 100. However, quality in the context of a PNG controls the quality of *compression* and not the quality of the *image*. Since PNG is a lossless format, compression only affects the size of the image and not its quality. This resulted in PNG images of a large size, with no compression at all. The same behavior could be observed when toDataURL is called with the image type explicitly set to `image/png`, without a quality argument. The expected behavior is only observed if toDataURL is called with the image type set to `image/png` and the quality set to 0, since this provides the highest level of compression. According to [§4.12.4.4](https://html.spec.whatwg.org/multipage/scripting.html#serialising-bitmaps-to-a-file) of the HTML5 spec the quality argument should only apply to images of type `image/jpeg`, and if quality is not provided the user-agent should use a default value. This means that the spec was being violated, since the quality was set to 100 regardless of the image type.

`ImageBufferQt` uses [`QPixmap::save`](http://doc.qt.io/qt-5/qpixmap.html#save-2) to save the canvas data to a buffer. The quality argument to `QPixmap::save` can be a value from -1 to 100, with -1 being a sentinel value that tells `QPixmap::save` to use the default quality for that type. The fix was to consider the quality argument (along with the associated sanity-checks) only if the image type is `image/jpeg`; otherwise quality is set to -1. This change results in PNG images being encoded to properly-sized base64 strings.