Update 2000-01-01-render.md
Updated with extra information based on the source code and my testing.
GIF export does not work (using PhantomJS 2.0 on Windows 7) - Qt documentation lists the GIF format as read-only, requiring a plugin to be able to write it. BMP and PPM export work without problems. XBM works, but looks terrible. XPM appears to work, but the files will not load.