Don't force Content-Type header on POST request with empty body
QT will still complain about that though if there is a query string in the request because the patch here: https://codereview.qt-project.org/#/c/99686/2/src/network/access/qhttpnetworkrequest.cpp has a flaw. It should not have the OR request.d->url.hasQuery() at the end, that's not what Content-Type is for.