Update selector in useragent example to prevent TypeError
The DOM structure of the page that is retrieved by useragent.js
has changed since the example was last updated, and the selector
for the user agent value no longer exists. This causes a TypeError
to be output to the user when running the example. This updates
the example to use a selector that exists on the page.

https://github.com/ariya/phantomjs/issues/15392